refactor(test): split use_ability + content-engine loader tests under cap
Tier 3 of #529. Both test files exceeded the 700-line hard cap. Pure
mechanical move — no test logic changed, every #[test] preserved.

use_ability/tests.rs (1091) -> tests/ dir with shared fixtures in
mod.rs and themed siblings: gating, target_validity, holster_queue,
auto_cycle, weapon_grant (26 tests preserved).

content-engine loader/tests.rs (900) -> tests/ dir with themed
siblings: chain_loading, trigger_conversion, action_conversion,
condition_conversion (41 tests preserved).
